transfer reducer to context
